+++ src/firewallfabrik/platforms/iptables/_policy_compiler.py
@@ -68,6 +68,7 @@ from firewallfabrik.compiler.processors._service import (
     SeparateUserServices,
     VerifyCustomServices,
     VerifyIcmpTypes,
+    VerifyIpProtocols,
     VerifyPortRanges,
 )
 from firewallfabrik.core.objects import (
@@ -536,6 +537,7 @@ class PolicyCompiler_ipt(PolicyCompiler):
         self.add(VerifyCustomServices('verify custom services'))
         self.add(VerifyPortRanges('verify port ranges'))
         self.add(VerifyIcmpTypes('verify ICMP types'))
+        self.add(VerifyIpProtocols('verify IP protocols'))
         self.add(VerifyAddressRanges('verify address ranges'))
         self.add(VerifyAddresses('verify addresses'))
         self.add(VerifyMacAddresses('verify MAC addresses'))