+++ src/firewallfabrik/platforms/nftables/_nat_compiler.py
@@ -52,6 +52,7 @@ from firewallfabrik.compiler.processors._service import (
     SeparateTCPWithFlags,
     VerifyCustomServices,
     VerifyIcmpTypes,
+    VerifyIpProtocols,
     VerifyPortRanges,
 )
 from firewallfabrik.core.objects import (
@@ -320,6 +321,7 @@ class NATCompiler_nft(NATCompiler):
         self.add(VerifyCustomServices('verify custom services'))
         self.add(VerifyPortRanges('verify port ranges'))
         self.add(VerifyIcmpTypes('verify ICMP types'))
+        self.add(VerifyIpProtocols('verify IP protocols'))
         self.add(VerifyAddressRanges('verify address ranges'))
         self.add(VerifyAddresses('verify addresses'))
         self.add(VerifyMacAddresses('verify MAC addresses'))